fix: fix issue with poptip circular dependencies
