+++ packages/vrender-components/src/poptip/poptip.ts
@@ -28,7 +28,7 @@ import {
 import { AbstractComponent } from '../core/base';
 import type { BackgroundAttributes, ComponentOptions } from '../interface';
 import type { PopTipAttributes } from './type';
-import { loadPoptip, loadPoptipComponent } from './register';
+import { loadPoptipComponent } from './register';
 
 const _tBounds = new AABBBounds();
