+++ packages/vrender-components/src/poptip/register.ts
@@ -1,34 +1,8 @@
-import {
-  container,
-  InteractiveSubRenderContribution,
-  AutoEnablePlugins,
-  ContainerModule
-} from '@visactor/vrender-core';
+import { container } from '@visactor/vrender-core';
 import { merge } from '@visactor/vutils';
-import { PopTipRenderContribution } from './contribution';
-import { PopTipPlugin, PopTipForClipedTextPlugin } from './poptip-plugin';
 import type { PopTipAttributes } from './type';
 import { DEFAULT_THEME, theme } from './theme';
 import { registerGroup, registerRect, registerSymbol, registerWrapText } from '@visactor/vrender-kits';
-
-export const popTipModule = new ContainerModule((bind, unbind, isBound, rebind) => {
-  if (!isBound(PopTipRenderContribution)) {
-    bind(PopTipRenderContribution).toSelf().inSingletonScope();
-    bind(InteractiveSubRenderContribution).toService(PopTipRenderContribution);
-  }
-  if (!isBound(PopTipPlugin)) {
-    bind(PopTipPlugin).toSelf();
-    bind(AutoEnablePlugins).toService(PopTipPlugin);
-  }
-  if (!isBound(PopTipForClipedTextPlugin)) {
-    bind(PopTipForClipedTextPlugin).toSelf();
-    bind(AutoEnablePlugins).toService(PopTipForClipedTextPlugin);
-  }
-});
-
-export function loadPoptip() {
-  container.load(popTipModule);
-}
 
 export function loadPoptipComponent() {
   registerGroup();
